Issue #11 - Fixing adding child spec exception
