+++ src/main/groovy/com/xebialabs/gradle/xldeploy/XlDeployPlugin.groovy
@@ -28,13 +28,23 @@ class XlDeployPlugin implements Plugin<Project> {
 
     addExtensions(project)
 
+
     DarTask dar = project.tasks.create(DAR_TASK_NAME, DarTask.class)
-    dar.configure {
+
+
+    DarConfigurationTask darConfiguration = project.tasks.create("darConfiguration", DarConfigurationTask.class)
+    darConfiguration.configure {
       group = BasePlugin.BUILD_GROUP
       description = "Create a DAR package"
+      darTask = dar
     }
+
+    dar.configure {
+      dependsOn darConfiguration
+    }
+
     ArchivePublishArtifact darArtifact = new ArchivePublishArtifact(dar)
-    project.configurations.create(DarTask.DAR_CONFIGURATION_NAME).artifacts.add(darArtifact)
+    project.configurations.create(DarConfigurationTask.DAR_CONFIGURATION_NAME).artifacts.add(darArtifact)
 
     DeployTask deploy = project.tasks.create(DEPLOY_TASK_NAME, DeployTask.class)
     deploy.configure {
